user-service/requirements.txt
flask
flask_smorest
flask_marshmallow
flask_sqlalchemy
flask_cors
Flask-JWT-Extended
sqlalchemy_utils
marshmallow_sqlalchemy
